% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cil.R
\name{cil}
\alias{cil}
\title{Confidence interval length}
\usage{
cil(x, alpha)
}
\arguments{
\item{x}{numeric vector}

\item{alpha}{level of significance}
}
\value{
confidence interval length at the given level of significance
}
\description{
The cil() function calculates the confidence interval parameters of a numeric vector.
It takes in as input the numeric vector and the level of significance desired for the
computation. As an output, it gives the population mean’s confidence interval lower and
upper bounds, as well as a number that represents the length of the confidence interval,
i.e., the difference between the upper bound and the lower bound of the confidence interval.
}
\examples{

\dontrun{vecx <- rnorm(100)
cil(vecx, 0.05)}

}
