% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{Relevance<-}
\alias{Relevance<-}
\alias{Relevance<-,Variant,character-method}
\title{Set clinical relevance}
\usage{
Relevance(object) <- value

\S4method{Relevance}{Variant,character}(object) <- value
}
\arguments{
\item{object}{An object of class Variant}

\item{value}{The clinical relevance}
}
\value{
The modified object
}
\description{
Set the clinical relevance of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "old_id",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
Relevance(snv) <- "new_rel"
Relevance(snv)
}
