% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\name{cormat}
\alias{cormat}
\title{Correlation matrix:
Calculates and returns the correlation matrix of a dataframe. Specifically designed to work with large dataset using parallelization tools.}
\usage{
cormat(df)
}
\arguments{
\item{df}{dataframe for which a correlation matrix is to be calculated}
}
\value{
Correlation matrix
}
\description{
Correlation matrix:
Calculates and returns the correlation matrix of a dataframe. Specifically designed to work with large dataset using parallelization tools.
}
\examples{
\dontrun{cormat(data.frame(matrix(rnorm(100), ncol = 5)))
closeAllConnections()}
}
