% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{Deletion-class}
\alias{Deletion-class}
\title{Class Deletion: Deletion Variant}
\description{
S4 class for a deletion variant
}
\section{Slots}{

\describe{
\item{\code{lengthDeleted}}{Integer. The length of deleted sequence.}
}}

\examples{
gr <- GenomicRanges::GRanges("chr13", IRanges::IRanges(32914438, 32914438),
 strand = "+")
del <- Deletion(
  VariantID = "DEL1",
  GenomicPos = gr,
  ReferenceAllele = "G",
  MutatedAllele = "-",
  lengthDeleted = as.integer(1)
)
isFrameshift(del)
}
