% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructors.R
\name{Deletion}
\alias{Deletion}
\title{Constructor for class Deletion}
\usage{
Deletion(
  VariantID,
  GenomicPos,
  ReferenceAllele,
  MutatedAllele = "-",
  lengthDeleted = NA_integer_,
  Exon = NA_character_,
  GeneSymbol = NA_character_,
  GeneType = NA_character_,
  MutationConsequence = NA_character_,
  ClinicalRel = NA_character_,
  EnsemblID = NA_character_,
  ProteinVariation = NA_character_
)
}
\arguments{
\item{VariantID}{Character. Unique ID of the variant.}

\item{GenomicPos}{\code{GRanges} object for the position of the variant.}

\item{ReferenceAllele}{Character. The allele in the reference genome.}

\item{MutatedAllele}{Character. The alternative or mutated allele.}

\item{lengthDeleted}{Integer. The length of deleted sequence.}

\item{Exon}{Character. The exon where the variant is located.}

\item{GeneSymbol}{Character. Gene name (optional).}

\item{GeneType}{Character. Type of gene e.g. "protein_coding" (optional).}

\item{MutationConsequence}{Character. Effect in the phenotype (optional).}

\item{ClinicalRel}{Character. Clinical annotation (optional).}

\item{EnsemblID}{Character. Transcript from Ensembl (optional).}

\item{ProteinVariation}{Character. HGVS protein annotation (optional).}
}
\value{
class \code{Deletion} object
gr <- GenomicRanges::GRanges("chr13", IRanges::IRanges(32914438, 32914438),
strand = "+")
del <- Deletion(
VariantID = "DEL1",
GenomicPos = gr,
ReferenceAllele = "G",
lengthDeleted = as.integer(1)
)
isFrameshift(del)
}
\description{
Constructor to create an object of class Deletion
}
