% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{describe}
\alias{describe}
\alias{describe,Variant-method}
\alias{describe,SNV-method}
\alias{describe,DNV-method}
\alias{describe,ONV-method}
\alias{describe,Insertion-method}
\alias{describe,Deletion-method}
\title{describe}
\usage{
describe(object)

\S4method{describe}{Variant}(object)

\S4method{describe}{SNV}(object)

\S4method{describe}{DNV}(object)

\S4method{describe}{ONV}(object)

\S4method{describe}{Insertion}(object)

\S4method{describe}{Deletion}(object)
}
\arguments{
\item{object}{An object of class Variant}
}
\value{
A character string
}
\description{
Returns a description of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100), strand = "+")
snv <- SNV(
  VariantID = "SNV001",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
describe(snv)

dnv <- DNV(
  VariantID = "DNV001",
  GenomicPos = gr,
  ReferenceAllele = "C",
  MutatedAllele = "T",
  validated = TRUE
)
describe(dnv)
}
