% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{DNV-class}
\alias{DNV-class}
\title{Class DNV: De Novo Variant}
\description{
S4 class for a de novo variant
}
\section{Slots}{

\describe{
\item{\code{validated}}{Logical. DNV has been confirmed}
}}

\examples{
gr <- GenomicRanges::GRanges("chr13", IRanges::IRanges(32914438, 32914438),
 strand = "+")
dnv <- DNV(
  VariantID = "DNV1",
  GenomicPos = gr,
  ReferenceAllele = "G",
  MutatedAllele = "T",
  validated = FALSE
)
}
