% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{EnsemblID<-}
\alias{EnsemblID<-}
\alias{EnsemblID<-,Variant,character-method}
\title{Set Ensembl ID}
\usage{
EnsemblID(object) <- value

\S4method{EnsemblID}{Variant,character}(object) <- value
}
\arguments{
\item{object}{An object of class Variant}

\item{value}{The Ensembl ID}
}
\value{
The modified object
}
\description{
Set the Ensembl ID of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "old_id",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
EnsemblID(snv) <- "new_id"
EnsemblID(snv)
}
