% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{EnsemblID}
\alias{EnsemblID}
\alias{EnsemblID,Variant-method}
\title{Get Ensembl ID}
\usage{
EnsemblID(object)

\S4method{EnsemblID}{Variant}(object)
}
\arguments{
\item{object}{An object of class Variant}
}
\value{
A character string
}
\description{
Returns the ensembl ID of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100), strand = "+")
snv <- SNV(
  VariantID = "SNV001",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
EnsemblID(snv)
}
