% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{Variant}
\format{
list with 11 S4 objects of different types
\describe{
\item{snv1, snv2, snv3}{Objects of class \code{SNV}}
\item{dnv1, dnv2, dnv3}{Objects of class \code{DNV}}
\item{onv1, onv2, onv3}{Objects of class \code{ONV}}
\item{insertion1, insertion2}{Objects of class \code{Insertion}}
\item{deletion1, deletion2}{Objects of class \code{Deletion}}}
}
\usage{
data(example)
}
\description{
A list containing different Variants
}
\examples{
data(example)
example$snv1
example$snv2
}
\keyword{datasets}
