% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{Exon<-}
\alias{Exon<-}
\alias{Exon<-,Variant,character-method}
\title{Set Exon}
\usage{
Exon(object) <- value

\S4method{Exon}{Variant,character}(object) <- value
}
\arguments{
\item{object}{An object of class Variant}

\item{value}{The exon to set}
}
\value{
The modifed object
}
\description{
Set the exon of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "variantid",
  Exon = "old_exon",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
Exon(snv) <- "new_exon"
Exon(snv)
}
