% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{Exon}
\alias{Exon}
\alias{Exon,Variant-method}
\title{Get Exon}
\usage{
Exon(object)

\S4method{Exon}{Variant}(object)
}
\arguments{
\item{object}{An object of class Variant}
}
\value{
A character string
}
\description{
Returns the exon of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "SNV001",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
Exon(snv)
}
