% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_trans.R
\name{fisher_trans}
\alias{fisher_trans}
\title{Fisher's Transformation.
Transforms the correlation "r" values into z-scores using the Fisher's transformation formula.}
\usage{
fisher_trans(r)
}
\arguments{
\item{r}{is the correlation coefficient to be transformed into a z score}
}
\value{
z-score
}
\description{
Fisher's Transformation.
Transforms the correlation "r" values into z-scores using the Fisher's transformation formula.
}
\examples{
\dontrun{fisher_trans(0.89)}
}
