% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{GeneSymbol<-}
\alias{GeneSymbol<-}
\alias{GeneSymbol<-,Variant,character-method}
\title{Set Gene symbol}
\usage{
GeneSymbol(object) <- value

\S4method{GeneSymbol}{Variant,character}(object) <- value
}
\arguments{
\item{object}{An object of class Variant}

\item{value}{Gene symbol to set}
}
\value{
The modifed object
}
\description{
Set the gene symbol of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "old_id",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  GeneSymbol = "symbol",
  transition = TRUE
)
GeneSymbol(snv) <- "new_symbol"
GeneSymbol(snv)
}
