% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{GeneSymbol}
\alias{GeneSymbol}
\alias{GeneSymbol,Variant-method}
\title{Get Gene Symbol}
\usage{
GeneSymbol(object)

\S4method{GeneSymbol}{Variant}(object)
}
\arguments{
\item{object}{An object of class Variant}
}
\value{
A character string
}
\description{
Returns the gene symbol of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "SNV001",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
GeneSymbol(snv)
}
