% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{Position}
\alias{Position}
\alias{Position,Variant-method}
\title{Get Position of variant}
\usage{
Position(object)

\S4method{Position}{Variant}(object)
}
\arguments{
\item{object}{An object of class GRanges}
}
\value{
A character string
}
\description{
Returns the position of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100), strand = "+")
snv <- SNV(
  VariantID = "SNV001",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
Position(snv)
}
