% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{Insertion-class}
\alias{Insertion-class}
\title{Class Insertion: Insertion Variant}
\description{
S4 class for an insertion variant
}
\section{Slots}{

\describe{
\item{\code{lengthInserted}}{Integer. The length of inserted sequence.}
}}

\examples{
gr <- GenomicRanges::GRanges("chr13", IRanges::IRanges(32914438, 32914438),
 strand = "+")
ins <- Insertion(
  VariantID = "INS1",
  GenomicPos = gr,
  ReferenceAllele = "-",
  MutatedAllele = "G",
  lengthInserted = as.integer(1)
)
isFrameshift(ins)
}
