% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructors.R
\name{Insertion}
\alias{Insertion}
\title{Constructor for class Insertion}
\usage{
Insertion(
  VariantID,
  GenomicPos,
  ReferenceAllele = "-",
  MutatedAllele,
  lengthInserted = NA_integer_,
  Exon = NA_character_,
  GeneSymbol = NA_character_,
  GeneType = NA_character_,
  MutationConsequence = NA_character_,
  ClinicalRel = NA_character_,
  EnsemblID = NA_character_,
  ProteinVariation = NA_character_
)
}
\arguments{
\item{VariantID}{Character. Unique ID of the variant.}

\item{GenomicPos}{\code{GRanges} object for the position of the variant.}

\item{ReferenceAllele}{Character. The allele in the reference genome.}

\item{MutatedAllele}{Character. The alternative or mutated allele.}

\item{lengthInserted}{Integer. The length of inserted sequence.}

\item{Exon}{Character. The exon where the variant is located.}

\item{GeneSymbol}{Character. Gene name (optional).}

\item{GeneType}{Character. Type of gene e.g. "protein_coding" (optional).}

\item{MutationConsequence}{Character. Effect in the phenotype (optional).}

\item{ClinicalRel}{Character. Clinical annotation (optional).}

\item{EnsemblID}{Character. Transcript from Ensembl (optional).}

\item{ProteinVariation}{Character. HGVS protein annotation (optional).}
}
\value{
class \code{Insertion} object
}
\description{
Constructor to create an object of class Insertion
}
\examples{
gr <- GenomicRanges::GRanges("chr13", IRanges::IRanges(32914438, 32914438),
 strand = "+")
ins <- Insertion(
  VariantID = "INS1",
  GenomicPos = gr,
  MutatedAllele = "G",
  lengthInserted = as.integer(1)
)
isFrameshift(ins)
}
