% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{isFrameshift}
\alias{isFrameshift}
\alias{isFrameshift,Insertion-method}
\alias{isFrameshift,Deletion-method}
\title{Is Frameshift}
\usage{
isFrameshift(object)

\S4method{isFrameshift}{Insertion}(object)

\S4method{isFrameshift}{Deletion}(object)
}
\arguments{
\item{object}{An object of class Insertion}
}
\value{
A logical value
}
\description{
Returns TRUE if is a frameshift mutation
}
\examples{
gr <- GenomicRanges::GRanges("chr13", IRanges::IRanges(32914438, 32914438),
 strand = "+")
ins <- Insertion(
  VariantID = "INS1",
  GenomicPos = gr,
  ReferenceAllele = "-",
  MutatedAllele = "T",
  lengthInserted = as.integer(1)
)
isFrameshift(ins)
gr <- GenomicRanges::GRanges("chr13", IRanges::IRanges(32914438, 32914438),
 strand = "+")
del <- Deletion(
  VariantID = "DEL1",
  GenomicPos = gr,
  ReferenceAllele = "G",
  MutatedAllele = "-",
  lengthDeleted = as.integer(1)
)
isFrameshift(del)
}
