% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{isTransition}
\alias{isTransition}
\alias{isTransition,SNV-method}
\title{Is Transition}
\usage{
isTransition(object)

\S4method{isTransition}{SNV}(object)
}
\arguments{
\item{object}{An object of class SNV}
}
\value{
A logical value
}
\description{
Returns TRUE if transition
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(VariantID = "id", GenomicPos = gr, ReferenceAllele = "A",
 MutatedAllele = "G", transition = TRUE)
isTransition(snv)
}
