% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{MutatedAllele<-}
\alias{MutatedAllele<-}
\alias{MutatedAllele<-,Variant,character-method}
\title{Set mutated allele}
\usage{
MutatedAllele(object) <- value

\S4method{MutatedAllele}{Variant,character}(object) <- value
}
\arguments{
\item{object}{An object of class Variant}

\item{value}{The mutated allele}
}
\value{
The modified object
}
\description{
Set the mutated allele of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "variantid",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
MutatedAllele(snv) <- "new_mutated"
MutatedAllele(snv)
}
