% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{MutatedAllele}
\alias{MutatedAllele}
\alias{MutatedAllele,Variant-method}
\title{Get Mutated Allele}
\usage{
MutatedAllele(object)

\S4method{MutatedAllele}{Variant}(object)
}
\arguments{
\item{object}{An object of class Variant}
}
\value{
A character string
}
\description{
Returns the mutated allele the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "SNV001",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
MutatedAllele(snv)
}
