% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{MutationConsequence<-}
\alias{MutationConsequence<-}
\alias{MutationConsequence<-,Variant,character-method}
\title{Set mutation consequence}
\usage{
MutationConsequence(object) <- value

\S4method{MutationConsequence}{Variant,character}(object) <- value
}
\arguments{
\item{object}{An object of class Variant}

\item{value}{The mutation consequence}
}
\value{
The modified object
}
\description{
Set the mutation consequence of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "old_id",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
MutationConsequence(snv) <- "new_consequence"
MutationConsequence(snv)
}
