% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{MutationConsequence}
\alias{MutationConsequence}
\alias{MutationConsequence,Variant-method}
\title{Get Consequence}
\usage{
MutationConsequence(object)

\S4method{MutationConsequence}{Variant}(object)
}
\arguments{
\item{object}{An object of class Variant}
}
\value{
A character string
}
\description{
Returns the consequence of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "SNV001",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
MutationConsequence(snv)
}
