% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_overlap.R
\name{no_overlap}
\alias{no_overlap}
\title{Check for CI overlap}
\usage{
no_overlap(x, y, alpha)
}
\arguments{
\item{x}{vector x}

\item{y}{vector y}

\item{alpha}{level of significance alpha}
}
\value{
indices where no confidence interval overlap exists
}
\description{
The no_overlap() function is used to determine if two numeric vectors have overlapping
confidence intervals of their population means using a desired level of significance.
It takes as input two vectors and the level of significance and returns a Boolean value
TRUE if no overlap exists and FALSE if overlap exists.
}
\examples{
vecx <- rnorm(100)
vecy <- rnorm(100)
no_overlap(vecx, vecy, 0.05)

}
