% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{ONV-class}
\alias{ONV-class}
\title{Class ONV: Other Nucleotide Variant}
\description{
S4 class for an other nucleotide variant
}
\section{Slots}{

\describe{
\item{\code{complexity}}{Character. Type of other variant}
}}

\examples{
gr <- GenomicRanges::GRanges("chr22", IRanges::IRanges(29091840, 29091840),
 strand = "+")
onv <- ONV(
  VariantID = "ONV001",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  complexity = "translocation"
)
describe(onv)
VariantID(onv)
}
