% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructors.R
\name{ONV}
\alias{ONV}
\title{Constructor for class ONV}
\usage{
ONV(
  VariantID,
  GenomicPos,
  ReferenceAllele,
  MutatedAllele,
  complexity = NA_character_,
  Exon = NA_character_,
  GeneSymbol = NA_character_,
  GeneType = NA_character_,
  MutationConsequence = NA_character_,
  ClinicalRel = NA_character_,
  EnsemblID = NA_character_,
  ProteinVariation = NA_character_
)
}
\arguments{
\item{VariantID}{Character. Unique ID of the variant.}

\item{GenomicPos}{\code{GRanges} object for the position of the variant.}

\item{ReferenceAllele}{Character. The allele in the reference genome.}

\item{MutatedAllele}{Character. The alternative or mutated allele.}

\item{complexity}{Character. The type of complex mutation.}

\item{Exon}{Character. The exon where the variant is located.}

\item{GeneSymbol}{Character. Gene name (optional).}

\item{GeneType}{Character. Type of gene e.g. "protein_coding" (optional).}

\item{MutationConsequence}{Character. Effect in the phenotype (optional).}

\item{ClinicalRel}{Character. Clinical annotation (optional).}

\item{EnsemblID}{Character. Transcript from Ensembl (optional).}

\item{ProteinVariation}{Character. HGVS protein annotation (optional).}
}
\value{
class \code{ONV} object
}
\description{
Constructor to create an object of class ONV
}
\examples{
gr <- GenomicRanges::GRanges("chr22", IRanges::IRanges(29091840, 29091840),
 strand = "+")
onv <- ONV(
  VariantID = "ONV001",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  complexity = "translocation"
)
describe(onv)
VariantID(onv)
}
