% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_overlap.R
\name{percent_overlap}
\alias{percent_overlap}
\title{Confidence Interval Overlap Percentage}
\usage{
percent_overlap(x, y, alpha)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{alpha}{level of significance}
}
\value{
percentage overlap between the confidence intervals of the two numeric vectors
}
\description{
The percent_overlap() function is used to calculate the overlap percentage between
two numeric vectors. The function uses equation 3 described in chapter 2.
The function takes as input two numeric vectors and a desired level of significance and it
outputs the percentage overlap between the confidence intervals of population means of the
input vectors. If there’s no overlap of the confidence intervals, e.g., the upper bound of
the confidence interval of one vector is lower than the lower bound of the confidence
interval of the other vector, a negative percent overlap would be displayed.
}
\examples{

\dontrun{vecx <- rnorm(100)
vecy <- rnorm(100)
percent_overlap(vecx, vecy, 0.05)}


}
