% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{ProteinVariation<-}
\alias{ProteinVariation<-}
\alias{ProteinVariation<-,Variant,character-method}
\title{Set protein variation}
\usage{
ProteinVariation(object) <- value

\S4method{ProteinVariation}{Variant,character}(object) <- value
}
\arguments{
\item{object}{An object of class Variant}

\item{value}{The protein variation}
}
\value{
The modified object
}
\description{
Set the protein variation of the variant
}
\details{
Set the protein variation of a Variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100), strand = "+")
snv <- SNV(
  VariantID = "old_id",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
ProteinVariation(snv) <- "new_variation"
ProteinVariation(snv)
}
