% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{ProteinVariation}
\alias{ProteinVariation}
\alias{ProteinVariation,Variant-method}
\title{Get Protein Variation}
\usage{
ProteinVariation(object)

\S4method{ProteinVariation}{Variant}(object)
}
\arguments{
\item{object}{An object of class Variant}
}
\value{
A character string
}
\description{
Returns the protein variation of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "SNV001",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
ProteinVariation(snv)
}
