% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{ReferenceAllele<-}
\alias{ReferenceAllele<-}
\alias{ReferenceAllele<-,Variant,character-method}
\title{Set reference allele}
\usage{
ReferenceAllele(object) <- value

\S4method{ReferenceAllele}{Variant,character}(object) <- value
}
\arguments{
\item{object}{An object of class Variant}

\item{value}{Reference allele}
}
\value{
The modifed object
}
\description{
Set the reference allele of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "variantid",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
ReferenceAllele(snv) <- "new_reference"
ReferenceAllele(snv)
}
