% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{ReferenceAllele}
\alias{ReferenceAllele}
\alias{ReferenceAllele,Variant-method}
\title{Get Reference Allele}
\usage{
ReferenceAllele(object)

\S4method{ReferenceAllele}{Variant}(object)
}
\arguments{
\item{object}{An object of class Variant}
}
\value{
A character string
}
\description{
Returns the reference allele the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "SNV001",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
ReferenceAllele(snv)
}
