% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{Relevance}
\alias{Relevance}
\alias{Relevance,Variant-method}
\title{Get Clinical Relevance}
\usage{
Relevance(object)

\S4method{Relevance}{Variant}(object)
}
\arguments{
\item{object}{An object of class Variant}
}
\value{
A character string
}
\description{
Returns the clinical relevance of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "SNV001",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
Relevance(snv)
}
