% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{SNV-class}
\alias{SNV-class}
\title{Class SNV: Single Nucleotide Variant}
\description{
S4 class for a single nucleotide variant
}
\section{Slots}{

\describe{
\item{\code{transition}}{Logical. Substitution is a transition.}
}}

\examples{
gr <- GenomicRanges::GRanges("chr13", IRanges::IRanges(32914438, 32914438),
 strand = "+")
snv <- SNV(
  VariantID = "SNV1",
  GenomicPos = gr,
  ReferenceAllele = "G",
  MutatedAllele = "A",
)
}
