% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructors.R
\name{SNV}
\alias{SNV}
\title{Constructor for class SNV}
\usage{
SNV(
  VariantID,
  GenomicPos,
  ReferenceAllele,
  MutatedAllele,
  transition = NA,
  Exon = NA_character_,
  GeneSymbol = NA_character_,
  GeneType = NA_character_,
  MutationConsequence = NA_character_,
  ClinicalRel = NA_character_,
  EnsemblID = NA_character_,
  ProteinVariation = NA_character_
)
}
\arguments{
\item{VariantID}{Character. Unique ID of the variant.}

\item{GenomicPos}{\code{GRanges} object for the position of the variant.}

\item{ReferenceAllele}{Character. The allele in the reference genome.}

\item{MutatedAllele}{Character. The alternative or mutated allele.}

\item{transition}{Logical. TRUE when transition (A<->G, C<->T).}

\item{Exon}{Character. The exon where the variant is located.}

\item{GeneSymbol}{Character. Gene name (optional).}

\item{GeneType}{Character. Type of gene e.g. "protein_coding" (optional).}

\item{MutationConsequence}{Character. Effect in the phenotype (optional).}

\item{ClinicalRel}{Character. Clinical annotation (optional).}

\item{EnsemblID}{Character. Transcript from Ensembl (optional).}

\item{ProteinVariation}{Character. HGVS protein annotation (optional).}
}
\value{
class \code{SNV} object
}
\description{
Constructor to create object of class SNV
}
\examples{
gr <- GenomicRanges::GRanges("chr13", IRanges::IRanges(32914438, 32914438),
 strand = "+")
snv <- SNV(
  VariantID = "SNV1",
  GenomicPos = gr,
  ReferenceAllele = "G",
  MutatedAllele = "A",
)
}
