% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{Variant-class}
\alias{Variant-class}
\title{Class Variant: Generic Variant}
\description{
S4 class for a generic variant
}
\section{Slots}{

\describe{
\item{\code{VariantID}}{Character. ID of variant.}

\item{\code{GenomicPos}}{GRanges object. Genomic position.}

\item{\code{Exon}}{Character. Exons involved.}

\item{\code{ReferenceAllele}}{Character. Allele Reference.}

\item{\code{MutatedAllele}}{Character. Mutated allele.}

\item{\code{GeneSymbol}}{Character. Gene Symbol.}

\item{\code{GeneType}}{Character. The type of gene.}

\item{\code{MutationConsequence}}{Character. The biological consequence of variant.}

\item{\code{ClinicalRel}}{Character. The clinical relevance.}

\item{\code{EnsemblID}}{Character. The Ensembl ID.}

\item{\code{ProteinVariation}}{Character. The variation of protein.}
}}

\examples{
gr <- GenomicRanges::GRanges("chr1", 
IRanges::IRanges(100, 100), strand = "+")
v <- Variant(
  VariantID = "v1",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "T"
)
VariantID(v)
}
