\name{variant-package}
\alias{variant-package}
\alias{variant}
\docType{package}
\title{
\packageTitle{variant}
}
\description{
\packageDescription{variant}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{variant}
\packageIndices{variant}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{variant}

Maintainer: \packageMaintainer{variant}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
