% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructors.R
\name{Variant}
\alias{Variant}
\title{Constructor for Class Variant}
\usage{
Variant(
  VariantID,
  GenomicPos,
  ReferenceAllele,
  MutatedAllele,
  Exon = NA_character_,
  GeneSymbol = NA_character_,
  GeneType = NA_character_,
  MutationConsequence = NA_character_,
  ClinicalRel = NA_character_,
  EnsemblID = NA_character_,
  ProteinVariation = NA_character_
)
}
\arguments{
\item{VariantID}{Character. Unique ID of the variant.}

\item{GenomicPos}{\code{GRanges} object for the position of the variant.}

\item{ReferenceAllele}{Character. The allele in the reference genome.}

\item{MutatedAllele}{Character. The alternative or mutated allele.}

\item{Exon}{Character. The exon where the variant is located.}

\item{GeneSymbol}{Character. Gene name (optional).}

\item{GeneType}{Character. Type of gene e.g. "protein_coding" (optional).}

\item{MutationConsequence}{Character. Effect in the phenotype (optional).}

\item{ClinicalRel}{Character. Clinical annotation (optional).}

\item{EnsemblID}{Character. Transcript from Ensembl (optional).}

\item{ProteinVariation}{Character. HGVS protein annotation (optional).}
}
\value{
class \code{Variant} object
}
\description{
Function to create object of class Variant.
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
v <- Variant(
  VariantID = "v1",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "T"
)
VariantID(v)
}
