% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{VariantID<-}
\alias{VariantID<-}
\alias{VariantID<-,Variant,character-method}
\title{Set Variant ID}
\usage{
VariantID(object) <- value

\S4method{VariantID}{Variant,character}(object) <- value
}
\arguments{
\item{object}{An object of class Variant}

\item{value}{The ID to insert}
}
\value{
The modified object
}
\description{
Set the unique ID of the variant
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100),
 strand = "+")
snv <- SNV(
  VariantID = "old_id",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
VariantID(snv) <- "new_id"
VariantID(snv)
}
