% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{VariantID}
\alias{VariantID}
\alias{VariantID,Variant-method}
\title{Get Variant ID}
\usage{
VariantID(object)

\S4method{VariantID}{Variant}(object)
}
\arguments{
\item{object}{An object of class Variant}
}
\value{
A character string
}
\description{
Returns the unique ID of the variant
}
\details{
Get Variant ID
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100), strand = "+")
snv <- SNV(
  VariantID = "SNV001",
  GenomicPos = gr,
  ReferenceAllele = "A",
  MutatedAllele = "G",
  transition = TRUE
)
VariantID(snv)
}
