% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cil.R
\name{cil}
\alias{cil}
\title{Confidence interval length}
\usage{
cil(x, alpha)
}
\arguments{
\item{x}{numeric vector}

\item{alpha}{level of significance}
}
\value{
confidence interval length at the given level of significance
}
\description{
Confidence interval length
}
\examples{

\dontrun{vecx <- rnorm(100)
cil(vecx, 0.05)}

}
