% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_check.R
\name{df_check}
\alias{df_check}
\title{Dataframe check}
\usage{
df_check(df1, df2)
}
\arguments{
\item{df1}{First dataframe}

\item{df2}{Second dataframe}
}
\value{
Returns an error message if the two dataframes do not have the same number of columns or if the elements in both dataframes are not numeric.
}
\description{
Dataframe check
}
\examples{
\dontrun{df1 <- data.frame(matrix(rnorm(100), ncol = 5))
df2 <- data.frame(matrix(rnorm(100), ncol = 5))
df_check(df1, df2)}
}
