% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_overlap.R
\name{no_overlap}
\alias{no_overlap}
\title{Check for CI overlap}
\usage{
no_overlap(x, y, alpha)
}
\arguments{
\item{x}{vector x}

\item{y}{vector y}

\item{alpha}{level of significance alpha}
}
\value{
indices where no confidence interval overlap exists
}
\description{
Check for CI overlap
}
\examples{
vecx <- rnorm(100)
vecy <- rnorm(100)
no_overlap(vecx, vecy, 0.05)

}
