% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_overlap.R
\name{percent_overlap}
\alias{percent_overlap}
\title{Confidence Interval Overlap}
\usage{
percent_overlap(x, y, alpha)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{alpha}{level of significance}
}
\value{
percentage overlap between the confidence intervals of the two numeric vectors
}
\description{
Confidence Interval Overlap
}
\examples{

\dontrun{vecx <- rnorm(100)
vecy <- rnorm(100)
percent_overlap(vecx, vecy, 0.05)}


}
