% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_signal.R
\name{variant_signal}
\alias{variant_signal}
\title{Variant Signal
Detects the columns that changed significantly from one state in dataset 1 to another state in dataset 2.}
\usage{
variant_signal(df1, df2, alpha)
}
\arguments{
\item{df1}{dataframe 1}

\item{df2}{dataframe 2}

\item{alpha}{level of significance e.g. 0.05}
}
\value{
list of columns that show significant change
}
\description{
Variant Signal
Detects the columns that changed significantly from one state in dataset 1 to another state in dataset 2.
}
\examples{

\dontrun{df1 <- data.frame(matrix(rnorm(100), ncol = 5))
df2 <- data.frame(matrix(rnorm(100), ncol = 5))
variant_signal(df1, df2, 0.05)}


}
